Generate client for jobs
