+++ apps/provisioning/pkg/apis/provisioning/v0alpha1/webhook.go
@@ -1,0 +1,20 @@
+package v0alpha1
+
+import metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+type WebhookResponse struct {
+	metav1.TypeMeta `json:",inline"`
+
+	// HTTP Status code
+	// 200 implies that the payload was understood but nothing is required
+	// 202 implies that an async job has been scheduled to handle the request
+	Code int `json:"code,omitempty"`
+
+	// Optional message
+	Message string `json:"added,omitempty"`
+
+	// Jobs to be processed
+	// When the response is 202 (Accepted) the queued jobs will be returned
+	Job *JobSpec `json:"job,omitempty"`
+}
