+++ pkg/tests/apis/openapi_snapshots/provisioning.grafana.app-v0alpha1.json
@@ -2803,7 +2803,7 @@
         }
       },
       "com.github.grafana.grafana.apps.provisioning.pkg.apis.provisioning.v0alpha1.Job": {
-        "description": "The repository name and type are stored as labels",
+        "description": "When this code is changed, make sure to update the code generation. As of writing, this can be done via the hack dir in the root of the repo: ./hack/update-codegen.sh provisioning If you've opened the generated files in this dir at some point in VSCode, you may also have to re-open them to clear errors. The repository name and type are stored as labels",
         "type": "object",
         "properties": {
           "apiVersion": {
@@ -4360,7 +4360,7 @@
         }
       },
       "com.github.grafana.grafana.pkg.apis.provisioning.v0alpha1.Job": {
-        "description": "The repository name and type are stored as labels",
+        "description": "When this code is changed, make sure to update the code generation. As of writing, this can be done via the hack dir in the root of the repo: ./hack/update-codegen.sh provisioning If you've opened the generated files in this dir at some point in VSCode, you may also have to re-open them to clear errors. The repository name and type are stored as labels",
         "type": "object",
         "properties": {
           "apiVersion": {
